+++ lang/src/org/partiql/lang/ast/StatementToExprNode.kt
@@ -84,36 +84,20 @@ private class StatementTransformer(val ion: IonSystem) {
             is PartiqlAst.Expr.Gte -> NAry(NAryOp.GTE, operands.toExprNodeList(), metas)
             is PartiqlAst.Expr.Lt -> NAry(NAryOp.LT, operands.toExprNodeList(), metas)
             is PartiqlAst.Expr.Lte -> NAry(NAryOp.LTE, operands.toExprNodeList(), metas)
-            is PartiqlAst.Expr.Union ->
-                NAry(
-                    when (setq) {
-                        is PartiqlAst.SetQuantifier.Distinct -> NAryOp.UNION
-                        is PartiqlAst.SetQuantifier.All -> NAryOp.UNION_ALL
-                    },
-                    operands.toExprNodeList(),
-                    metas
-                )
-            is PartiqlAst.Expr.Intersect ->
-                NAry(
-                    when (setq) {
-                        is PartiqlAst.SetQuantifier.Distinct -> NAryOp.INTERSECT
-                        is PartiqlAst.SetQuantifier.All -> NAryOp.INTERSECT_ALL
-                    },
-                    operands.toExprNodeList(),
-                    metas
-                )
-            is PartiqlAst.Expr.Except ->
-                NAry(
-                    when (setq) {
-                        is PartiqlAst.SetQuantifier.Distinct -> NAryOp.EXCEPT
-                        is PartiqlAst.SetQuantifier.All -> NAryOp.EXCEPT_ALL
-                    },
-                    operands.toExprNodeList(),
-                    metas
-                )
-            is PartiqlAst.Expr.OuterUnion,
-            is PartiqlAst.Expr.OuterIntersect,
-            is PartiqlAst.Expr.OuterExcept -> error("$this node has no representation in prior ASTs.")
+            is PartiqlAst.Expr.BagOp -> {
+                if (op is PartiqlAst.BagOpType.OuterUnion ||
+                    op is PartiqlAst.BagOpType.OuterIntersect ||
+                    op is PartiqlAst.BagOpType.OuterExcept
+                ) {
+                    error("$this node has no representation in prior ASTs.")
+                }
+                var id = op.javaClass.simpleName.toUpperCase()
+                if (quantifier is PartiqlAst.SetQuantifier.All) {
+                    id += "_ALL"
+                }
+                val nAryOp = NAryOp.valueOf(id)
+                NAry(nAryOp, operands.toExprNodeList(), metas)
+            }
             is PartiqlAst.Expr.Like -> NAry(NAryOp.LIKE, listOfNotNull(value.toExprNode(), pattern.toExprNode(), escape?.toExprNode()), metas)
             is PartiqlAst.Expr.Between -> NAry(NAryOp.BETWEEN, listOf(value.toExprNode(), from.toExprNode(), to.toExprNode()), metas)
             is PartiqlAst.Expr.InCollection -> NAry(NAryOp.IN, operands.toExprNodeList(), metas)